Move `database_match_type` to its own file and add conversion extensions
